+++ NEWS.md
@@ -1,5 +1,6 @@
 # FSA 0.4.19 ongoing
 * Modified some tests to check whether the suggested package was installed.
+* `capHistConver()`: Modified.  Added a m-array object for when more than two sampling events are present.b
 * `mrClosed()`:  Modified.  Fixed bugs around printing of CI type with Schnabel and the ignoring of `conf.level=` with Schnabel.
 * `plot.mrClosed()`:  Modified.  Changed axis labels as the expressions did not print with some fonts and devices.